Removed redundant casts

+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/TrackTablePanel.java
@@ -59,7 +59,7 @@ public final class TrackTablePanel extends EntityTablePanel {
                                                                                                       final SwingEntityEditModel editModel,
                                                                                                       final Integer initialValue) {
       final MinutesSecondsPanelValue minutesSecondsPanelValue = new MinutesSecondsPanelValue();
-      minutesSecondsPanelValue.set((Integer) initialValue);
+      minutesSecondsPanelValue.set(initialValue);
 
       return minutesSecondsPanelValue;
     }
